Replace RenderingManager with OnGUIEvery()
This was based on advice from a KSP forum post by Sarbian explaining how
to convert plugins to KSP 1.1 [0]. While the code now compiles, nothing
happens in-game when the toolbar button is hovered over and/or clicked,
so evidently something is still broken.

Don't know if the documentation fix is technically correct, but it's at
least more correct than it was earlier. Hopefully.

    [0]: https://forum.kerbalspaceprogram.com/index.php?/topic/135179-info-on-how-to-convert-your-plugin-to-ksp-11/