Call base OnGUIEvery() in MonoBehaviourWindowPlus
MonoBehaviourWindow takes care of drawing the info window, as well as
deciding whether the window should be drawn. The code actually invoking
DrawGUI() is in OnGUIEvery().

It appears that the RealScience info window is an instance of
MonoBehaviourWindowPlus, which derives from MonoBehaviourWindow. The
former overrides OnGUIEvery(), so DrawGUI() was never being called,
which would explain why the window was not showing up.